FIX: Error when opening filters after switching search type
This fixes an issue where you get an error if you do a topic search, then switch to chat messages and click the advanced filters button `this.args.uiState` is null.

This also removes the `currentRefinements` data attribute. It's not currently being used in CSS and conditionally showing the refinements would be better done in the JS instead of CSS anyways.